+++ packages/uba/onUserLogin.js
@@ -10,10 +10,10 @@ if (ubaAuditPresent) {
 /**
  * Checking of user IP and device fingerpriont based on settings from `uba_advSecurity`
  * @param {THTTPRequest} req
+ * @return {{enabled: false}}{enabled: true, kmn: string, fpa: string}}
  */
 function checkAdvancedSecurity (req) {
   let advData
-
   try {
     advData = UB.Repository('uba_advSecurity')
       .attrs(['ID', 'allowedIP', 'refreshIP', 'fp', 'refreshFp', 'keyMediaName', 'refreshKeyMedia', 'mi_modifyDate'])
@@ -22,9 +22,17 @@ function checkAdvancedSecurity (req) {
   } catch (e) {
     // nothing to do - table uba_advSecurity not exists
     console.warn('Advanced security is disabled because table uba_advSecurity does not exists')
-    doCheckAdvancedSecurity = function () {}
+    doCheckAdvancedSecurity = function () {return {enabled: false}}
   }
-  if (!advData) return // no adv. settings for current user
+  let fp = ''
+  let fpa = ''
+  let urlParams = queryString.parse(req.decodedParameters)
+
+  if (!advData) return { // no adv. settings for current user
+    ebnabled: true,
+    kmn: urlParams.KMN || '',
+    fpa: urlParams.FPA || ''
+  }
   let updateParams = {}
   let needUpdate = false
   if (advData.refreshIP) {
@@ -33,11 +41,6 @@ function checkAdvancedSecurity (req) {
     needUpdate = true
   } else if (advData.allowedIP) {
     if (Session.callerIP !== advData.allowedIP) throw new Error('Allowed IP ' + advData.allowedIP + ' <> actual ' + Session.callerIP)
-  }
-  let fp = ''
-  let urlParams
-  if (advData.refreshFp || advData.fp || advData.refreshKeyMedia || advData.keyMediaName) { // fp required
-    urlParams = queryString.parse(req.parameters)
   }
 
   if (advData.refreshFp || advData.fp) { // fp required
@@ -72,6 +75,11 @@ function checkAdvancedSecurity (req) {
     })
     advStore.freeNative()
   }
+  return {
+    enabled: true,
+    kmn: urlParams.KMN || '',
+    fpa: urlParams.FPA || ''
+  }
 }
 
 let doCheckAdvancedSecurity = null // calculate later
@@ -89,12 +97,10 @@ function onUserLogin (req) {
 
   let userInfo = UB.Repository('uba_user').attrs('name').selectById(Session.userID)
   data.login = userInfo.name || Session.userID
-
   if (!doCheckAdvancedSecurity) {
-    doCheckAdvancedSecurity = App.domainInfo.has('uba_advSecurity') ? checkAdvancedSecurity : function () {}
+    doCheckAdvancedSecurity = App.domainInfo.has('uba_advSecurity') ? checkAdvancedSecurity : function () { return {enabled: false} }
   }
-  doCheckAdvancedSecurity(req)
-
+  let advCheckData = doCheckAdvancedSecurity(req)
   try {
     repo = UB.Repository('uba_userrole')
       .attrs(['ID', 'roleID.name', 'roleID'])
@@ -127,7 +133,8 @@ function onUserLogin (req) {
           actionUser: Session.userID,
           actionTime: new Date(),
           remoteIP: Session.callerIP,
-          targetUser: data.login,
+          targetUser: (advCheckData.enabled && advCheckData.kmn) ? advCheckData.kmn : data.login,
+          targetRole: (advCheckData.enabled && advCheckData.fpa) ? advCheckData.fpa.slice(0, 127) : '',
           fromValue: req.headers
         }
       })